Replaced double with float.

+++ FormulaRendering/FormulaRenderResult.cs
@@ -6,8 +6,8 @@ namespace WallpaperGenerator.FormulaRendering
     {
         public Size Size { get; private set; }
 
-        public double[] EvaluatedValuesBuffer { get; private set; }
-        public double[] ColorTranformedValuesBuffer { get; private set; }
+        public float[] EvaluatedValuesBuffer { get; private set; }
+        public float[] ColorTranformedValuesBuffer { get; private set; }
 
         public byte[] RedChannel { get; private set; }
         public byte[] GreenChannel { get; private set; }
@@ -17,8 +17,8 @@ namespace WallpaperGenerator.FormulaRendering
         {
             Size = size;
             int l = size.Width*size.Height;
-            EvaluatedValuesBuffer = new double[l];
-            ColorTranformedValuesBuffer = new double[l];
+            EvaluatedValuesBuffer = new float[l];
+            ColorTranformedValuesBuffer = new float[l];
             RedChannel = new byte[l];
             GreenChannel = new byte[l];
             BlueChannel = new byte[l];
